+++ src/dash-table/dash/DataTable.js
@@ -87,6 +87,7 @@ export const defaultProps = {
     data: [],
     columns: [],
     editable: false,
+    export_format: 'none',
     selected_cells: [],
     selected_rows: [],
     selected_row_ids: [],
@@ -116,6 +117,25 @@ export const propTypes = {
      * `name` and `id` are the only required parameters.
      */
     columns: PropTypes.arrayOf(PropTypes.exact({
+
+        /**
+         * If True, the user can clear the column by clicking on a little `Ø`
+         * button on the column.
+         * If there are merged, multi-header columns then you can choose
+         * which column header row to display the "Ø" in by
+         * supplying an array of booleans.
+         * For example, `[true, false]` will display the "Ø" on the first row,
+         * but not the second row.
+         * If the "Ø" appears on a merged column, then clicking on that button
+         * will clear *all* of the merged columns associated with it.
+         *
+         * Unlike `column.deletable`, this action does not remove the column(s)
+         * from the table. It only removed the associated entries from `data`.
+         */
+        clearable: PropTypes.oneOfType([
+            PropTypes.bool,
+            PropTypes.arrayOf(PropTypes.bool)
+        ]),
 
         /**
          * If True, the user can delete the column by clicking on a little `x`
@@ -415,6 +435,23 @@ export const propTypes = {
         row_id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
         column_id: PropTypes.string
     }),
+
+    /**
+     * Denotes the type of the export data file,
+     * Defaults to `'none'`
+     */
+    export_format: PropTypes.oneOf(['csv', 'xlsx', 'none']),
+
+    /**
+     * Denotes the format of the headers in the export data file.
+     * If `'none'`, there will be no header. If `'display'`, then the header
+     * of the data file will be be how it is currently displayed. Note that
+     * `'display'` is only supported for `'xlsx'` export_format and will behave
+     * like `'names'` for `'csv'` export format. If `'ids'` or `'names'`,
+     * then the headers of data file will be the column id or the column
+     * names, respectively
+     */
+    export_headers: PropTypes.oneOf(['none', 'ids', 'names', 'display']),
 
     /**
      * `fill_width` toggles between a set of CSS for two common behaviors: