+++ src/dash-table/derived/header/content.tsx
@@ -13,15 +13,42 @@ import {
     VisibleColumns,
     IVisibleColumn,
     SetProps,
-    TableAction
+    TableAction,
+    SetFilter
 } from 'dash-table/components/Table/props';
 import * as actions from 'dash-table/utils/actions';
+import { SingleColumnSyntaxTree } from 'dash-table/syntax-tree';
+import { clearColumnsFilter } from '../filter/map';
 
-function deleteColumn(column: IVisibleColumn, columns: VisibleColumns, columnRowIndex: any, setProps: SetProps, data: Data) {
-    return () => {
-        setProps(actions.deleteColumn(column, columns, columnRowIndex, data));
-    };
-}
+const doAction = (
+    action: (
+        column: IVisibleColumn,
+        columns: VisibleColumns,
+        columnRowIndex: any,
+        mergeDuplicateHeaders: boolean,
+        data: Data
+    ) => any,
+    column: IVisibleColumn,
+    columns: VisibleColumns,
+    columnRowIndex: any,
+    mergeDuplicateHeaders: boolean,
+    setFilter: SetFilter,
+    setProps: SetProps,
+    map: Map<string, SingleColumnSyntaxTree>,
+    data: Data
+) => () => {
+    setProps(action(column, columns, columnRowIndex, mergeDuplicateHeaders, data));
+
+    const affectedColumns: VisibleColumns = [];
+    R.forEach(id => {
+        const affectedColumn = columns.find(c => c.id === id);
+        if (affectedColumn) {
+            affectedColumns.push(affectedColumn);
+        }
+    }, actions.getAffectedColumns(column, columns, columnRowIndex, mergeDuplicateHeaders));
+
+    clearColumnsFilter(map, affectedColumns, setFilter);
+};
 
 function doSort(columnId: ColumnId, sortBy: SortBy, mode: SortMode, setProps: SetProps) {
     return () => {
@@ -55,9 +82,9 @@ function doSort(columnId: ColumnId, sortBy: SortBy, mode: SortMode, setProps: Se
     };
 }
 
-function editColumnName(column: IVisibleColumn, columns: VisibleColumns, columnRowIndex: any, setProps: SetProps) {
+function editColumnName(column: IVisibleColumn, columns: VisibleColumns, columnRowIndex: any, setProps: SetProps, mergeDuplicateHeaders: boolean) {
     return () => {
-        setProps(actions.editColumnName(column, columns, columnRowIndex));
+        setProps(actions.editColumnName(column, columns, columnRowIndex, mergeDuplicateHeaders));
     };
 }
 
@@ -79,15 +106,25 @@ function getSortingIcon(columnId: ColumnId, sortBy: SortBy) {
     }
 }
 
+function getColumnFlag(i: number, flag?: boolean | boolean[]): boolean {
+    return typeof flag === 'boolean' ?
+        flag :
+        !!flag && flag[i];
+}
+
 function getter(
     columns: VisibleColumns,
+    mergeDuplicateHeaders: boolean,
     data: Data,
     labelsAndIndices: R.KeyValuePair<any[], number[]>[],
+    map: Map<string, SingleColumnSyntaxTree>,
     sort_action: TableAction,
     mode: SortMode,
     sortBy: SortBy,
     paginationMode: TableAction,
-    setProps: SetProps
+    setFilter: SetFilter,
+    setProps: SetProps,
+    merge_duplicate_headers: boolean
 ): JSX.Element[][] {
     return R.addIndex<R.KeyValuePair<any[], number[]>, JSX.Element[]>(R.map)(
         ([labels, indices], headerRowIndex) => {
@@ -97,15 +134,9 @@ function getter(
                 columnIndex => {
                     const column = columns[columnIndex];
 
-                    const renamable: boolean = typeof column.renamable === 'boolean' ?
-                        column.renamable :
-                        !!column.renamable && column.renamable[headerRowIndex];
-
-                    const deletable = paginationMode !== TableAction.Custom && (
-                        typeof column.deletable === 'boolean' ?
-                            column.deletable :
-                            !!column.deletable && column.deletable[headerRowIndex]
-                    );
+                    const renamable = getColumnFlag(headerRowIndex, column.renamable);
+                    const clearable = paginationMode !== TableAction.Custom && getColumnFlag(headerRowIndex, column.clearable);
+                    const deletable = paginationMode !== TableAction.Custom && getColumnFlag(headerRowIndex, column.deletable);
 
                     return (<div>
                         {sort_action !== TableAction.None && isLastRow ?
@@ -121,24 +152,28 @@ function getter(
                         {renamable ?
                             (<span
                                 className='column-header--edit'
-                                onClick={editColumnName(column, columns, headerRowIndex, setProps)}
-                            >
-                                {`✎`}
-                            </span>) :
+                                onClick={editColumnName(column, columns, headerRowIndex, setProps, merge_duplicate_headers)}
+                            />) :
+                            ''
+                        }
+
+                        {clearable ?
+                            (<span
+                                className='column-header--clear'
+                                onClick={doAction(actions.clearColumn, column, columns, headerRowIndex, mergeDuplicateHeaders, setFilter, setProps, map, data)}
+                            />) :
                             ''
                         }
 
                         {deletable ?
                             (<span
                                 className='column-header--delete'
-                                onClick={deleteColumn(column, columns, headerRowIndex, setProps, data)}
-                            >
-                                {'×'}
-                            </span>) :
+                                onClick={doAction(actions.deleteColumn, column, columns, headerRowIndex, mergeDuplicateHeaders, setFilter, setProps, map, data)}
+                            />) :
                             ''
                         }
 
-                        <span>{labels[columnIndex]}</span>
+                        <span className='column-header-name'>{labels[columnIndex]}</span>
                     </div>);
                 },
                 indices