+++ src/dash-table/utils/actions.js
@@ -1,9 +1,8 @@
 import * as R from 'ramda';
+import getHeaderRows from 'dash-table/derived/header/headerRows';
 
-function getGroupedColumnIndices(column, columns, headerRowIndex) {
-    const columnIndex = columns.indexOf(column);
-
-    if (!column.name || (Array.isArray(column.name) && column.name.length < headerRowIndex)) {
+function getGroupedColumnIndices(column, columns, headerRowIndex, mergeDuplicateHeaders, columnIndex) {
+    if (!column.name || (Array.isArray(column.name) && column.name.length < headerRowIndex) || !mergeDuplicateHeaders) {
         return { groupIndexFirst: columnIndex, groupIndexLast: columnIndex };
     }
 
@@ -22,22 +21,38 @@ function getGroupedColumnIndices(column, columns, headerRowIndex) {
     return { groupIndexFirst: columnIndex, groupIndexLast: lastColumnIndex };
 }
 
-export function deleteColumn(column, columns, headerRowIndex, data) {
-    const {groupIndexFirst, groupIndexLast} = getGroupedColumnIndices(
-        column, columns, headerRowIndex
+export function getAffectedColumns(column, columns, headerRowIndex, mergeDuplicateHeaders) {
+    const { groupIndexFirst, groupIndexLast } = getGroupedColumnIndices(
+        column, columns, headerRowIndex, mergeDuplicateHeaders, columns.indexOf(column)
     );
-    const rejectedColumnIds = R.slice(
+
+    return R.slice(
         groupIndexFirst,
         groupIndexLast + 1,
         R.pluck('id', columns)
     );
+}
+
+export function clearColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data) {
+    const rejectedColumnIds = getAffectedColumns(column, columns, headerRowIndex, mergeDuplicateHeaders);
+
+    return {
+        data: R.map(R.omit(rejectedColumnIds), data)
+    };
+}
+
+export function deleteColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data) {
+    const {groupIndexFirst, groupIndexLast} = getGroupedColumnIndices(
+        column, columns, headerRowIndex, mergeDuplicateHeaders, columns.indexOf(column)
+    );
+
     return {
         columns: R.remove(
             groupIndexFirst,
             1 + groupIndexLast - groupIndexFirst,
             columns
         ),
-        data: R.map(R.omit(rejectedColumnIds), data),
+        ...clearColumn(column, columns, headerRowIndex, mergeDuplicateHeaders, data),
         // NOTE - We're just clearing these so that there aren't any
         // inconsistencies. In an ideal world, we would probably only
         // update them if they contained one of the columns that we're
@@ -53,23 +68,34 @@ export const clearSelection = {
     selected_cells: []
 };
 
-export function editColumnName(column, columns, headerRowIndex) {
+export function changeColumnHeader(column, columns, headerRowIndex, mergeDuplicateHeaders, newColumnName) {
+    let newColumns = columns;
+    const maxLength = getHeaderRows(newColumns);
+    const columnIndex = newColumns.findIndex(col => col.id === column.id);
+
+    if (typeof column.name === 'string' && maxLength > 1) {
+        const newColumnNames = Array(maxLength).fill(column.name);
+        const cloneColumn = R.mergeRight(column, {name: newColumnNames});
+        newColumns = newColumns.slice(0);
+        newColumns[columnIndex] = cloneColumn;
+    }
+
     const { groupIndexFirst, groupIndexLast } = getGroupedColumnIndices(
-        column, columns, headerRowIndex
+        column, newColumns, headerRowIndex, mergeDuplicateHeaders, columnIndex
     );
-    /* eslint no-alert: 0 */
-    const newColumnName = window.prompt('Enter a new column name');
-    let newColumns = R.clone(columns);
+
     R.range(groupIndexFirst, groupIndexLast + 1).map(i => {
         let namePath;
-        if (R.type(columns[i].name) === 'Array') {
+        if (R.type(newColumns[i].name) === 'Array') {
             namePath = [i, 'name', headerRowIndex];
-        } else {
-            namePath = [i, 'name'];
         }
         newColumns = R.set(R.lensPath(namePath), newColumnName, newColumns);
     });
-    return {
-        columns: newColumns
-    };
+
+    return { columns: newColumns} ;
+}
+
+export function editColumnName(column, columns, headerRowIndex, mergeDuplicateHeaders) {
+    const newColumnName = window.prompt('Enter a new column name');
+    return changeColumnHeader(column, columns, headerRowIndex, mergeDuplicateHeaders, newColumnName);
 }